requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
streamlit>=1.39.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
langsmith>=0.1.0
playwright>=1.44.0

# After pip install, also run: playwright install chromium
